Fix ignore pymemcache's get_many returns False value
In some case, pymemcache returns {'key': False}
if client does not found.

It raises exception at Django-Cachalot.

Django-Cachalot's author said get_many should not contain a key
that's not in the cache, as stated in Django.

So we try to imprements get_many method and ignore if False
were included.